Update Card.java
I think this might be a better improvement to the Card class which makes the Card class an algebraic product type of the enumerations, Suit, and Rank. This also eliminates the need for all the previous methods that were used to find the suit, color, number value, suit symbol UTF8 string, and number symbol. Plus, it reads better and is more intuitive. Of course, this new implementation will require some refactoring of those classes that reference the Card class.